added delete recipe on frontend

+++ assets/scripts/main.js
@@ -73,42 +73,25 @@ $(function () {
 		});
 	});
 
-	/*
 
-	 $('.delete-recipe').on('click', function (e) {
-	 e.preventDefault();
-	 $this = $(this);
+	$('.delete-recipe').on('click', function (e) {
+		e.preventDefault();
 
-	 var recipeId = $this.data('id');
+		if (window.confirm("Are you sure?")) {
+			$this = $(this);
 
-	 $.ajax({
-	 url    : '/recipes/' + recipeId,
-	 type   : 'DELETE',
-	 success: function (result) {
-	 // Do something with the result
-	 console.log(result);
-	 }
-	 });
-	 });
+			var url = $this.data('url');
 
-	 $('.update-recipe').on('click', function (e) {
-	 e.preventDefault();
-	 $this = $(this);
+			$.ajax({
+				url    : url,
+				type   : 'DELETE',
+				success: function (result) {
+					// Do something with the result
+					$this.parent().remove();
+				}
+			});
+		}
 
-	 var recipeId = $this.data('id'),
-	 updateTitle = $this.data('title');
-
-	 $.ajax({
-	 url    : '/recipes/' + recipeId,
-	 type   : 'PUT',
-	 data   : {
-	 'title': updateTitle
-	 },
-	 success: function (result) {
-	 // Do something with the result
-	 console.log(result);
-	 }
-	 });
-	 });
-	 */
+	});
+	
 });
